requirement.txt
autopep8
django==4.0
djangorestframework
requests
python-dotenv
gunicorn
django_heroku
pytest